backend/requirements.txt
flask
supabase
python-dotenv
requests
app